+++ lib/karafka/runner.rb
@@ -14,7 +14,8 @@ module Karafka
       listeners = Connection::ListenersBatch.new(jobs_queue)
 
       workers.each(&:async_call)
-      listeners.each(&:async_call)
+
+      App.config.internal.connection.manager.register(listeners)
 
       # We aggregate threads here for a supervised shutdown process
       Karafka::Server.workers = workers
@@ -56,29 +57,27 @@ module Karafka
     # @param listeners [Connection::ListenersBatch]
     def wait_actively(listeners)
       # Thread#join requires time in seconds, thus the conversion
-      join_timeout = Karafka::App.config.internal.join_timeout / 1_000.to_f
+      join_timeout = Karafka::App.config.internal.join_timeout / 1_000.0
 
       listeners.cycle do |listener|
         # Do not manage rebalances if we're done and shutting down
         break if Karafka::App.done?
+
+        next unless listener.active?
 
         # If join returns something else than nil, it means this listener is either stopped or
         # paused and its timeout will not kick in. In such cases we do not use its timeout because
         # it is not time reliable and could trigger the event too often
         next unless listener.join(join_timeout).nil?
 
-        Karafka.monitor.instrument(
-          'runner.join_timeout',
-          caller: self,
-          listeners: listeners
-        )
+        Karafka.monitor.instrument('runner.join_timeout', caller: self)
       end
 
-      # All the listener threads need to finish when shutdown is happening
+      # All the active listener threads need to finish when shutdown is happening
       # If we are here it means we're done and shutdown has started. During this period we no
       # longer tick from the runner but we still need to wait on all the listeners to finish all
       # the work
-      listeners.each(&:join)
+      listeners.active.each(&:join)
     end
   end
 end